pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "luna-bench"
version = "0.0.1a3"
description = "Benchmarking framework for classical, hybrid and quantum optimization algorithms"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.13"
authors = [
    { name = "Aqarios GmbH", email = "support@aqarios.com" },
]
keywords = [
    "aqarios",
    "luna",
    "quantum computing",
    "quantum optimization",
    "optimization",
    "benchmarking",
]
dependencies = [
    "dependency-injector>=4.48.2",
    "huey>=3.0.0",
    "luna-model>=0.6.0",
    "luna-quantum>=1.2.1",
    "peewee>=4.0.0",
    "pip-licenses>=5.5.1",
    "pydantic-settings>=2.11.0",
    "pyrefly>=0.45.0",
    "returns[compatible-mypy]>=0.27.0",
    "rich>=14.1.0",
]

[project.urls]
Homepage = "https://aqarios.com"
Documentation = "https://docs.aqarios.com"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pandas-stubs>=3.0.0.260204",
    "networkx>=3.6.1",
    "pyscipopt>=6.0.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-env >=1.6.0",
    "ruff>=0.15.0",
    "scipy-stubs>=1.17.0.2",
    "tox>=4.30.3",
    "types-networkx>=3.6.1.20260210",
    "types-peewee>=4.0.0.20260223",
    "types-seaborn>=0.13.2.20251221",
    "pyscipopt>=6.0.1",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.0",
    "networkx>=3.6.1",
    "scipy>=1.17.0",
]

[project.optional-dependencies]
opt-metrics = [
    "pyscipopt>=6.0.1",
]
plot = [
    "matplotlib>=3.10.8",
    "seaborn>=0.13.0",
]
qubo-features = [
    "networkx>=3.6.1",
    "scipy>=1.17.0",
]

[tool.setuptools]
packages = ["luna_bench"]
